Apply Factory pattern to canvas worker
