+++ src/components/Editor/DrawingBoard/Canvas/Worker/Worker.ts
@@ -1,0 +1,19 @@
+import { ToolType } from 'features/boardSlices';
+import { Point } from 'features/docSlices';
+import { LineOption } from '../line';
+
+type Options = LineOption;
+
+interface Worker {
+  type: ToolType;
+
+  mousedown: (point: Point, options: Options) => void;
+
+  mousemove: (point: Point) => void;
+
+  mouseup: () => void;
+
+  flushTask: () => void;
+}
+
+export default Worker;
