+++ src/components/Editor/DrawingBoard/Canvas/Worker/index.ts
@@ -1,0 +1,5 @@
+export type { default as Worker } from './Worker';
+export { default as LineWorker } from './LineWorker';
+export { default as RectWorker } from './RectWorker';
+export { default as EraserWorker } from './EraserWorker';
+export { default as SelectorWorker } from './SelectorWorker';
